Fix memory crash: Zero-copy numpy conversion

+++ training/train_minigpt.py
@@ -75,8 +75,8 @@ print(f"Val tokens: {len(val_data):,}")
 def get_batch(split):
     dataset = train_data if split == 'train' else val_data
     ix = torch.randint(len(dataset) - block_size, (batch_size,))
-    x = torch.stack([dataset[i:i+block_size] for i in ix])
-    y = torch.stack([dataset[i+1:i+block_size+1] for i in ix])
+    x = torch.stack([dataset[i:i+block_size] for i in ix]).long()
+    y = torch.stack([dataset[i+1:i+block_size+1] for i in ix]).long()
     return x.to(device), y.to(device)
 
 # Model
